refactor(session): split instance.rs into terminal/container submodules
Cleanup-pass finding #6 from the code-quality scan. No behavior changes;
all 1079 unit tests pass.

Before: src/session/instance.rs was 1,607 lines hosting one impl block
that mixed core lifecycle, paired-terminal management, and container
orchestration.

After:
  - src/session/instance/mod.rs (1,453 lines) — Instance + related struct
    definitions, core lifecycle (start/restart/kill/stop, status update,
    capture_output), and tests.
  - src/session/instance/terminal.rs (123 lines) — paired-terminal impl
    block: terminal_tmux_session, has_terminal, start_terminal,
    start_terminal_with_size, kill_terminal, container_terminal_*,
    apply_*_terminal_tmux_options.
  - src/session/instance/container.rs (65 lines) — container impl block:
    get_container_for_instance, container_workdir, build_container_config.

Rust allows multiple impl blocks across files for the same type, so the
split keeps the public Instance API unchanged.
