+++ src/session/instance/container.rs
@@ -1,0 +1,65 @@
+//! Container lifecycle for `Instance`: image pulls, container creation,
+//! and workspace-path computation for sandboxed sessions.
+
+use std::path::Path;
+
+use anyhow::Result;
+
+use crate::containers::{self, ContainerRuntimeInterface, DockerContainer};
+
+use super::super::container_config;
+use super::Instance;
+
+impl Instance {
+    pub fn get_container_for_instance(&mut self) -> Result<containers::DockerContainer> {
+        let sandbox = self
+            .sandbox_info
+            .as_ref()
+            .ok_or_else(|| anyhow::anyhow!("Cannot ensure container for non-sandboxed session"))?;
+
+        let image = &sandbox.image;
+        let container = DockerContainer::new(&self.id, image);
+
+        if container.is_running()? {
+            container_config::refresh_agent_configs();
+            return Ok(container);
+        }
+
+        if container.exists()? {
+            container_config::refresh_agent_configs();
+            container.start()?;
+            return Ok(container);
+        }
+
+        // Ensure image is available (always pulls to get latest)
+        let runtime = containers::get_container_runtime();
+        runtime.ensure_image(image)?;
+
+        let config = self.build_container_config()?;
+        let container_id = container.create(&config)?;
+
+        if let Some(ref mut sandbox) = self.sandbox_info {
+            sandbox.container_id = Some(container_id);
+        }
+
+        Ok(container)
+    }
+
+    /// Get the container working directory for this instance.
+    pub fn container_workdir(&self) -> String {
+        container_config::compute_volume_paths(Path::new(&self.project_path), &self.project_path)
+            .map(|(_, wd)| wd)
+            .unwrap_or_else(|_| "/workspace".to_string())
+    }
+
+    fn build_container_config(&self) -> Result<crate::containers::ContainerConfig> {
+        container_config::build_container_config(
+            &self.project_path,
+            self.sandbox_info.as_ref().unwrap(),
+            &self.tool,
+            self.is_yolo_mode(),
+            &self.id,
+            self.workspace_info.as_ref(),
+        )
+    }
+}
