+++ src/session/instance/terminal.rs
@@ -1,0 +1,123 @@
+//! Paired-terminal lifecycle (host + container terminals) for `Instance`.
+//!
+//! Methods are attached to `Instance` via an `impl` block in this module;
+//! the struct itself lives in `super::mod`.
+
+use anyhow::Result;
+
+use crate::tmux;
+
+use super::super::environment::build_docker_env_args;
+use super::{Instance, TerminalInfo};
+
+impl Instance {
+    pub fn terminal_tmux_session(&self) -> Result<tmux::TerminalSession> {
+        tmux::TerminalSession::new(&self.id, &self.title)
+    }
+
+    pub fn has_terminal(&self) -> bool {
+        self.terminal_info
+            .as_ref()
+            .map(|t| t.created)
+            .unwrap_or(false)
+    }
+
+    pub fn start_terminal(&mut self) -> Result<()> {
+        self.start_terminal_with_size(None)
+    }
+
+    pub fn start_terminal_with_size(&mut self, size: Option<(u16, u16)>) -> Result<()> {
+        let session = self.terminal_tmux_session()?;
+
+        let is_new = !session.exists();
+        if is_new {
+            session.create_with_size(&self.project_path, None, size)?;
+        }
+
+        // Apply all configured tmux options to terminal sessions too
+        if is_new {
+            self.apply_terminal_tmux_options();
+        }
+
+        self.terminal_info = Some(TerminalInfo { created: true });
+
+        Ok(())
+    }
+
+    pub fn kill_terminal(&self) -> Result<()> {
+        let session = self.terminal_tmux_session()?;
+        if session.exists() {
+            session.kill()?;
+        }
+        Ok(())
+    }
+
+    pub fn container_terminal_tmux_session(&self) -> Result<tmux::ContainerTerminalSession> {
+        tmux::ContainerTerminalSession::new(&self.id, &self.title)
+    }
+
+    pub fn has_container_terminal(&self) -> bool {
+        self.container_terminal_tmux_session()
+            .map(|s| s.exists())
+            .unwrap_or(false)
+    }
+
+    pub fn start_container_terminal_with_size(&mut self, size: Option<(u16, u16)>) -> Result<()> {
+        if !self.is_sandboxed() {
+            anyhow::bail!("Cannot create container terminal for non-sandboxed session");
+        }
+
+        let container = self.get_container_for_instance()?;
+        let sandbox = self.sandbox_info.as_ref().unwrap();
+
+        let env_info = build_docker_env_args(sandbox, std::path::Path::new(&self.project_path));
+        let env_part = if env_info.docker_args.is_empty() {
+            String::new()
+        } else {
+            format!("{} ", env_info.docker_args)
+        };
+
+        // Get workspace path inside container (handles bare repo worktrees correctly)
+        let container_workdir = self.container_workdir();
+
+        let cmd = container.exec_command(
+            Some(&format!("-w {} {}", container_workdir, env_part)),
+            "/bin/bash",
+        );
+
+        // If there are secret env vars, prepend shell exports and use `exec`
+        // so the outer shell (whose argv briefly contains the export values)
+        // is replaced immediately, keeping secrets out of long-lived process argv.
+        let session_cmd = if env_info.exports.is_empty() {
+            cmd
+        } else {
+            let exports = env_info.exports.join("; ");
+            format!("{}; exec {}", exports, cmd)
+        };
+
+        let session = self.container_terminal_tmux_session()?;
+        let is_new = !session.exists();
+        if is_new {
+            session.create_with_size(&self.project_path, Some(&session_cmd), size)?;
+            self.apply_container_terminal_tmux_options();
+        }
+
+        Ok(())
+    }
+
+    pub fn kill_container_terminal(&self) -> Result<()> {
+        let session = self.container_terminal_tmux_session()?;
+        if session.exists() {
+            session.kill()?;
+        }
+        Ok(())
+    }
+    fn apply_container_terminal_tmux_options(&self) {
+        let name = tmux::ContainerTerminalSession::generate_name(&self.id, &self.title);
+        self.apply_session_tmux_options(&name, &format!("{} (container)", self.title));
+    }
+    fn apply_terminal_tmux_options(&self) {
+        let name = tmux::TerminalSession::generate_name(&self.id, &self.title);
+        self.apply_session_tmux_options(&name, &format!("{} (terminal)", self.title));
+    }
+}
